Refactor app, update README and .travis.yml
