+++ .travis.yml
@@ -1,7 +1,7 @@
 language: python
 python: '3.4'
 install: pip install -r requirements.txt
-script: python manage.py test
+script: python manage.py test tests
 notifications:
   slack:
     secure: BS9g3jKOF2cAGra2NBOoUMcvJFX3hP8YUIwtjhR63kB5BshVXghZdiTaat4XeH8U96KhixWfLsxYJnRI1VxM2/vFwAEIxWP3zdDbOdavVsfzX7KCi+W6MSErjbz7tKLyERzEa6/EdXe+0fe0/2Mzm96hXennKDjh1SSvcDhlHm/5S2LyMGqAB2w4b2UYO54EZ9hcejSvQs1L2d2YfIkUqXuY+vX9pNTMBFvmcogcfT99syhK0JNcbjhqlR2hLFZL2NCs7VNHtvTVjvQa9+33MXvy/8stXNBocO9f0H0v+Izc1bvDdkn0zJjZsXmiKdXD3ku+7Xye0tOiDLzi50W/dQZaI3BSeSn5XLIM1Yk84PZxnrLdhGRsjITNe/qidbq3/PTOUu9su6pvQuU9lwFwV91pzyDhI6IEMOu8REJam5s4AOsE3pwkKn//vhG59InMfx6A8lw54qKgoXVB1N8KvgzkiW5BPcJcz6m8r62Lu2WUJWS5qco1zG+Dt4FKizI98uyQ7IoHBwFwQ26oEAhyz+3VNYuCkF/aXEvvttzUAAvB32ANy7jsFbXy2bP5T/oAVSXiBZEtwwMdJK3WJXyazOQxJm2Jni6VZQ4ovE5H7k7MO2mPaVOyFXmb8+h2CmKKEOob7idbHd0y4JBOtk087+BrE9EKPWabXxsfgsMXjgo=
