+++ app/admin.py
@@ -3,8 +3,7 @@
 ## =============================================================================
 
 from django.contrib import admin
-from .models import Movie
-from .models import Tweet
+from app.models.models import *
 
 # Register your models here.
 admin.site.register(Movie)
