+++ app/urls.py
@@ -3,7 +3,7 @@
 ## =============================================================================
 
 from django.conf.urls import url
-from . import views
+from .views import views
 
 urlpatterns = [
     url(r'^index', views.index, name='index'),
