+++ app/views/views.py
@@ -5,7 +5,9 @@
 
 from django.shortcuts import render
 from django.http import HttpResponse
-from .models import *
+from app.models.models import *
+from app.models.helpers import *
+from app.views.helpers import *
 import datetime
 from django.utils import timezone
 from django.contrib import messages
@@ -78,7 +80,6 @@ def results(request):
         data_to_render = {'error_message': ''}
         sum_scores = 0
         num_nonzero = 1
-
 
         ## If no search term was given, pick a random one
         if not search_term or search_term == '':
@@ -155,28 +156,7 @@ def results(request):
             data_to_render['error_message'] = 'Sorry, we couldn\'t find enough tweets about that movie for analysis.'
             return render(request, 'error.html', data_to_render)
 
-        ## Chart sentiment scores of tweets
-        overall_score = get_overall_sentiment_score(clean_tweets)
-        polarity = get_polarity(clean_tweets)
-        negative_data, positive_data, neutral_count = create_chart_datasets(clean_tweets)
-        tweets_to_display = get_tweets_to_display(clean_tweets)
-
-        ## Save our new sentiment data to the db
-        save_new_sentiment(overall_score, movie)
-
-        ## Prepare data to render on results page
-        data_to_render = {  'form'          : QueryForm(request.POST),
-                            'tweets'        : tweets_to_display,
-                            'movie'         : movie,
-                            'overall_score' : overall_score,
-                            'polarity'      : polarity,
-                            'new_form'      : QueryForm(),
-                            'negative_data' : negative_data, # data for the chart
-                            'positive_data' : positive_data, # data for the chart
-                            'negative_count': len(negative_data), # data for the chart
-                            'positive_count': len(positive_data), # data for the chart
-                            'neutral_count' : neutral_count # data for the chart
-                        }
+        data_to_render = prepare_data_for_render(request, clean_tweets, movie)
         return render(request, 'data.html', data_to_render)
 
     ## Otherwise return METHOD NOT ALLOWED
@@ -185,116 +165,6 @@ def results(request):
 
 ## =============================================================================
 
+## Respond to request for about page
 def about(request):
     return render(request, 'about.html')
-
-## =============================================================================
-
-## Create a chart that displays the sentiment scores for the tweets associated with a movie
-def create_chart_datasets(clean_tweets):
-    negative_data = []
-    positive_data = []
-    neutral_count = 0
-
-    for tweet in clean_tweets:
-        score = tweet.sentiment_score
-
-        if score != 0:
-            data = {
-                        'y': round(abs(score)*100,1),
-                        'x': str(tweet.created_at),
-                        'r': 5
-                    }
-
-            if score < 0:
-                negative_data.append(data)
-            else:
-                positive_data.append(data)
-        else:
-            neutral_count += 1
-
-    return negative_data, positive_data, neutral_count
-
-## =============================================================================
-
-## Returns a list of Tweet objects created from the given list of twitter.Status objects
-def get_clean_tweets(raw_tweets, movie_title):
-    clean_tweets = []
-    for raw_tweet in raw_tweets:
-        clean_tweet = Tweet().fillWithStatusObject(raw_tweet, movie_title)
-        if clean_tweet:
-            clean_tweets.append(clean_tweet)
-    return clean_tweets
-
-## =============================================================================
-
-## Returns a list of at most 10 Tweet objects whos sentiment scores are not 0
-def get_tweets_to_display(clean_tweets):
-    tweets_to_display = []
-    i = 0
-    while i < len(clean_tweets) and len(tweets_to_display) < 10:
-        if clean_tweets[i].sentiment_score != 0:
-            clean_tweets[i].sentiment_score = round((clean_tweets[i].sentiment_score+1)*5, 1)
-            tweets_to_display.append(clean_tweets[i])
-        i += 1
-
-    return tweets_to_display
-
-## =============================================================================
-
-## Returns average sentiment score on a scale of 0 to 10, rounded to one decimal place
-def get_overall_sentiment_score(clean_tweets):
-    num_nonzero = 0
-    sum_scores = 0
-
-    for tweet in clean_tweets:
-        score = tweet.sentiment_score
-        if score != 0:
-            num_nonzero += 1
-            sum_scores += score
-
-    if num_nonzero == 0:
-        return 5.0
-    return round((sum_scores/num_nonzero+1)*5, 1)
-
-## =============================================================================
-
-## Returns polarity of scentiment scores as a percentage #TODO THIS IS TERRIBLE
-def get_polarity(clean_tweets):
-    positive_count = 0
-    negative_count = 0
-    positive_sum = 0
-    negative_sum = 0
-
-    for tweet in clean_tweets:
-        score = tweet.sentiment_score
-
-        if score > 0:
-            positive_count += 1
-            positive_sum += score
-        elif score < 0:
-            negative_count += 1
-            negative_sum += score
-
-    if positive_count != 0 and negative_count != 0:
-        return round((positive_sum/positive_count - negative_sum/negative_count)*50, 1)
-    else:
-        return 0
-
-## =============================================================================
-
-## Saves the given sentiment data to the db if it is not redundant
-def save_new_sentiment(overall_score, movie):
-    ## Find objects that are of the same movie and were made today
-    duplicate = Sentiment.objects.filter(imdbID = movie.imdbID, sentimentDate = timezone.now())
-
-    ## If there aren't duplicates create a new sentiment object for this movie
-    if not duplicate:
-        ## Populate the sentiment table with data for each movie with their scores at this time
-        MovieSentiment = Sentiment(Title = movie.Title, imdbID = movie.imdbID, sentimentDate = timezone.now() , sentimentScore = overall_score)
-
-        try:
-            ## Movie doesn't exist yet (in the db) with this date so save it into the db
-            MovieSentiment.save()
-        except:
-            print ("ERROR: couldn't save to sentiment table")
